pyproject.toml
[project]
name = "house-cli"
version = "0.1.0"
description = "CLI tool for searching houses across multiple platforms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "pyyaml>=6.0",
    "browser-cookie3>=0.19",
]

[project.scripts]
house = "house_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/house_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
